feat: implemented new endpoint to user and datasource creation
